Add is_zero_dimensional function
`is_zero_dimensional` decides for a Groebner basis whether the
generated ideal is zero-dimensional (or equivalently if the number of
solutions of the system is finite). It may also work for non-Groebner
bases but if `False` is returned, the ideal may still be
zero-dimensional.

It can be used as follows:
>>> from sympy.polys.fglm import is_zero_dimensional
>>> F = [x**2, x - y]  # y-axis intersected with line with slope 1
>>> is_zero_dimensional(F, x, y, order='lex')
False
>>> is_zero_dimensional(F, y, x, order='lex')  # lucky choice
True
>>> G = groebner(F, x, y, order='grlex')
>>> is_zero_dimensional(G, x, y, order='grlex')
True