added groebnerwalk
Still missing: Computing the next weight vector doesn't work yet and I'm
sure there are still a few bugs.

The basic idea of the Groebner walk is to have a Groebner basis
represented (?) by a weight vector and change it while moving along the
weight vector from the first order to the second.
This works because the weight vectors that represent one Groebner basis
are a cone in R^n and when crossing borders to another cone, only a
smaller problem has to be solved to obtain a Groebner basis for the new
cone.

     ____________________________
    /       \    \   /\          \
   / *Start  \    \ /  \ * Target \
   ¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯
The algorithm proceeds by finding the weightvector at the next border,
converting a Groebner basis of a smaller problem and lifting the smaller
GB to a GB of the original system in the new cone. Note: For every term
order there is a cone such that the Groebner basis w.r.t. this term
order is the Groebner basis represented by all points in the cone.