Added new solver for polynomial sytems
The new solver computes a Groebner basis (for now wrt to "grlex" - any
term order works!) and then finds the eigenvalues of the map F_i

   ``F_i(m) = normalform(m * x_i, G)``

from `K[x_1, ..., x_n]/(G)` into itself. The basis of `K[x_1, ...,
x_n]/(G)` is the set of monomials not divisible by any leading term of
``G``.
These eigenvalues are the i-th coordinates of the solutions of ``G``.

The problem with this approach is that computing the characteristic
polynomial of rather large matrices is slow. If we'd solve systems
numerically, this approach might be better because errors aren't
propagated. (If randomness is allowed, only a single matrix has to be
constructed).

So, this will probably not be very useful. It could be modified to
compute the minimal polynomial of an algebraic number.