Use FGLM for solving polynomial systems
Instead of computing the Groebner basis of a polynomial system wrt lex,
compute it wrt to grevlex and convert it afterwards. This is usually
much faster.
If GMPY is not available, the Groebner basis wrt lex is computed
directly (the linear algebra used in fglm is too slow without GMPY).